Use one Sender per PendingLogEntry instead of per batch
Instead of using a Sender per pool of entries, allocate one Sender per
pending entry. This is slightly less efficient as we're creating more
channels, but gives much more flexibility for how we handle individual
entries. More thorough benchmarking is needed, but from some quick tests
there was no noticeable difference. This change also allows us to stop
tracking the 'pool_index' of each log entry, and just transmit the final
index on each entry's channel.

This refactor will make it easier to only select a subset of the pending
entries each time sequencing occurs, so we can try to keep the tree as a
multiple of the full tile width (256), so we create fewer partial data
and level-0 tiles (#33).